Moved interfaces into types folder
