+++ client/src/pages/CV/components/Experience.tsx
@@ -1,11 +1,11 @@
 import List from "../../../components/List";
-import { IExperience } from "../ICvData";
+import { IExperience } from "../types/IExperience";
 import { formatInterval } from "../util/formatCvData";
 
 /**
  * Must pass IExperience object to the function
  * Returns experience component
- * @param {IExperience} experience of type IExperience containing data for one experience
+ * @param {IExperience} experience containing data for one experience
  * @returns div block with experience information structured
  */
 function Experience({ experience }: { experience: IExperience }) {
