+++ client/src/pages/CV/types/IEducation.ts
@@ -1,0 +1,5 @@
+export type IEducation = {
+    Degrees: string[];
+    School: string;
+    Interval: Date[];
+};
