+++ client/src/pages/CV/types/IExperience.ts
@@ -1,0 +1,11 @@
+export interface IExperience {
+    ID: number;
+    Type: string;
+    Name: string;
+    Team?: string;
+    Title: string;
+    Interval: Date[];
+    Location: string;
+    Description: string[];
+    Skills: string[];
+}
