+++ client/src/pages/CV/types/ILanguage.ts
@@ -1,0 +1,4 @@
+export type ILanguage = {
+    Name: string;
+    Level: string;
+};
